refactor: refactoring nhl widget
