+++ client/src/App.js
@@ -8,7 +8,7 @@ import DashBoard from "./pages/dashboard";
 import LoginPage from "./pages/login";
 import Cabinet from "./pages/cabinet";
 import Registration from "./pages/registration";
-import DetailPage from "./pages/dashboard/components/DetailPage";
+import DetailPage from "./pages/dashboard/components/Nhl/DetailPage";
 
 const App = () => {
   return (
