+++ client/src/pages/dashboard/components/Nhl/DetailPage/index.js
