+++ client/src/pages/dashboard/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import FavoriteTeams from "./components/FavoriteTeams";
+import FavoriteTeams from "./components/Nhl/FavoriteTeams";
 import DashSettings from "./components/Settings";
 import Show from "./components/Show";
